Fix #14932: Increase internal badge index size to avoid overflowing BadgeIDs. (#14933)
(cherry picked from commit 39f30692febbf288697e2dae16af550d741ce6ca)

# Conflicts:
#	src/newgrf_badge_type.h

+++ src/newgrf_badge_type.h
@@ -13,9 +13,9 @@
 #include "core/enum_type.hpp"
 #include "core/strong_typedef_type.hpp"
 
-struct BadgeIDTag : public StrongType::TypedefTraits<uint16_t, StrongType::Compare> {};
+struct BadgeIDTag : public StrongType::TypedefTraits<uint32_t, StrongType::Compare> {};
 using BadgeID = StrongType::Typedef<BadgeIDTag>;
-struct BadgeClassIDTag : public StrongType::TypedefTraits<uint16_t, StrongType::Compare> {};
+struct BadgeClassIDTag : public StrongType::TypedefTraits<uint32_t, StrongType::Compare> {};
 using BadgeClassID = StrongType::Typedef<BadgeClassIDTag>;
 
 template <> struct std::hash<BadgeClassID> {
